Move test_chunked_upload to test_lowlevel.py
